Add DataSourceMapSetter
