+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-boot-starter/src/main/java/org/apache/shardingsphere/driver/spring/boot/SpringBootConfiguration.java
@@ -17,17 +17,12 @@
 
 package org.apache.shardingsphere.driver.spring.boot;
 
-import com.google.common.base.Preconditions;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.driver.api.ShardingSphereDataSourceFactory;
+import org.apache.shardingsphere.driver.spring.boot.datasource.DataSourceMapSetter;
 import org.apache.shardingsphere.driver.spring.boot.prop.SpringBootPropertiesConfigurationProperties;
 import org.apache.shardingsphere.driver.spring.boot.rule.SpringBootRulesConfigurationProperties;
 import org.apache.shardingsphere.driver.spring.boot.rule.SpringBootRulesConfigurationYamlSwapper;
-import org.apache.shardingsphere.infra.exception.ShardingSphereException;
-import org.apache.shardingsphere.sharding.strategy.algorithm.sharding.inline.InlineExpressionParser;
-import org.apache.shardingsphere.driver.spring.boot.datasource.DataSourcePropertiesSetterHolder;
-import org.apache.shardingsphere.driver.spring.boot.util.DataSourceUtil;
-import org.apache.shardingsphere.driver.spring.boot.util.PropertyUtil;
 import org.apache.shardingsphere.driver.spring.transaction.ShardingTransactionTypeScanner;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -38,15 +33,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
-import org.springframework.core.env.StandardEnvironment;
-import org.springframework.jndi.JndiObjectFactoryBean;
 
-import javax.naming.NamingException;
 import javax.sql.DataSource;
 import java.sql.SQLException;
-import java.util.Collections;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -59,8 +49,6 @@ import java.util.Map;
 @AutoConfigureBefore(DataSourceAutoConfiguration.class)
 @RequiredArgsConstructor
 public class SpringBootConfiguration implements EnvironmentAware {
-    
-    private static final String JNDI_NAME = "jndi-name";
     
     private final SpringBootRulesConfigurationProperties rules;
     
@@ -93,44 +81,6 @@ public class SpringBootConfiguration implements EnvironmentAware {
     
     @Override
     public final void setEnvironment(final Environment environment) {
-        String prefix = "spring.shardingsphere.datasource.";
-        for (String each : getDataSourceNames(environment, prefix)) {
-            try {
-                dataSourceMap.put(each, getDataSource(environment, prefix, each));
-            } catch (final ReflectiveOperationException ex) {
-                throw new ShardingSphereException("Can't find data source type.", ex);
-            } catch (final NamingException ex) {
-                throw new ShardingSphereException("Can't find JNDI data source.", ex);
-            }
-        }
-    }
-    
-    private List<String> getDataSourceNames(final Environment environment, final String prefix) {
-        StandardEnvironment standardEnv = (StandardEnvironment) environment;
-        standardEnv.setIgnoreUnresolvableNestedPlaceholders(true);
-        return null == standardEnv.getProperty(prefix + "name")
-                ? new InlineExpressionParser(standardEnv.getProperty(prefix + "names")).splitAndEvaluate() : Collections.singletonList(standardEnv.getProperty(prefix + "name"));
-    }
-    
-    @SuppressWarnings("unchecked")
-    private DataSource getDataSource(final Environment environment, final String prefix, final String dataSourceName) throws ReflectiveOperationException, NamingException {
-        Map<String, Object> dataSourceProps = PropertyUtil.handle(environment, prefix + dataSourceName.trim(), Map.class);
-        Preconditions.checkState(!dataSourceProps.isEmpty(), "Wrong data source properties.");
-        if (dataSourceProps.containsKey(JNDI_NAME)) {
-            return getJndiDataSource(dataSourceProps.get(JNDI_NAME).toString());
-        }
-        DataSource result = DataSourceUtil.getDataSource(dataSourceProps.get("type").toString(), dataSourceProps);
-        DataSourcePropertiesSetterHolder.getDataSourcePropertiesSetterByType(dataSourceProps.get("type").toString()).ifPresent(
-            dataSourcePropertiesSetter -> dataSourcePropertiesSetter.propertiesSet(environment, prefix, dataSourceName, result));
-        return result;
-    }
-    
-    private DataSource getJndiDataSource(final String jndiName) throws NamingException {
-        JndiObjectFactoryBean bean = new JndiObjectFactoryBean();
-        bean.setResourceRef(true);
-        bean.setJndiName(jndiName);
-        bean.setProxyInterface(DataSource.class);
-        bean.afterPropertiesSet();
-        return (DataSource) bean.getObject();
+        dataSourceMap.putAll(DataSourceMapSetter.getDataSourceMap(environment));
     }
 }
