+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-orchestration-spring/shardingsphere-jdbc-orchestration-spring-boot-starter/src/main/java/org/apache/shardingsphere/driver/orchestration/spring/boot/OrchestrationSpringBootConfiguration.java
@@ -25,15 +25,11 @@ import org.apache.shardingsphere.driver.orchestration.spring.boot.common.SpringB
 import org.apache.shardingsphere.driver.orchestration.spring.boot.rule.LocalRulesCondition;
 import org.apache.shardingsphere.driver.orchestration.spring.boot.rule.OrchestrationSpringBootRulesConfigurationProperties;
 import org.apache.shardingsphere.driver.orchestration.spring.boot.rule.OrchestrationSpringBootRulesConfigurationYamlSwapper;
-import org.apache.shardingsphere.infra.exception.ShardingSphereException;
+import org.apache.shardingsphere.driver.spring.boot.datasource.DataSourceMapSetter;
 import org.apache.shardingsphere.orchestration.center.config.CenterConfiguration;
 import org.apache.shardingsphere.orchestration.center.config.OrchestrationConfiguration;
 import org.apache.shardingsphere.orchestration.center.yaml.config.YamlCenterRepositoryConfiguration;
 import org.apache.shardingsphere.orchestration.center.yaml.swapper.CenterRepositoryConfigurationYamlSwapper;
-import org.apache.shardingsphere.sharding.strategy.algorithm.sharding.inline.InlineExpressionParser;
-import org.apache.shardingsphere.driver.spring.boot.datasource.DataSourcePropertiesSetterHolder;
-import org.apache.shardingsphere.driver.spring.boot.util.DataSourceUtil;
-import org.apache.shardingsphere.driver.spring.boot.util.PropertyUtil;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -45,16 +41,12 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
-import org.springframework.core.env.StandardEnvironment;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.StringUtils;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
@@ -126,36 +118,6 @@ public class OrchestrationSpringBootConfiguration implements EnvironmentAware {
     
     @Override
     public final void setEnvironment(final Environment environment) {
-        String prefix = "spring.shardingsphere.datasource.";
-        for (String each : getDataSourceNames(environment, prefix)) {
-            try {
-                dataSourceMap.put(each, getDataSource(environment, prefix, each));
-            } catch (final ReflectiveOperationException ex) {
-                throw new ShardingSphereException("Can't find data source type.", ex);
-            }
-        }
-    }
-    
-    private List<String> getDataSourceNames(final Environment environment, final String prefix) {
-        StandardEnvironment standardEnv = (StandardEnvironment) environment;
-        standardEnv.setIgnoreUnresolvableNestedPlaceholders(true);
-        String dataSources = standardEnv.getProperty(prefix + "name");
-        if (StringUtils.isEmpty(dataSources)) {
-            dataSources = standardEnv.getProperty(prefix + "names");
-        }
-        if (StringUtils.isEmpty(dataSources)) {
-            return Collections.emptyList();
-        }
-        return new InlineExpressionParser(dataSources).splitAndEvaluate();
-    }
-    
-    @SuppressWarnings("unchecked")
-    private DataSource getDataSource(final Environment environment, final String prefix, final String dataSourceName) throws ReflectiveOperationException {
-        Map<String, Object> dataSourceProps = PropertyUtil.handle(environment, prefix + dataSourceName, Map.class);
-        Preconditions.checkState(!dataSourceProps.isEmpty(), String.format("Wrong datasource [%s] properties!", dataSourceName));
-        DataSource result = DataSourceUtil.getDataSource(dataSourceProps.get("type").toString(), dataSourceProps);
-        DataSourcePropertiesSetterHolder.getDataSourcePropertiesSetterByType(dataSourceProps.get("type").toString()).ifPresent(
-            dataSourcePropertiesSetter -> dataSourcePropertiesSetter.propertiesSet(environment, prefix, dataSourceName, result));
-        return result;
+        dataSourceMap.putAll(DataSourceMapSetter.getDataSourceMap(environment));
     }
 }
