+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-spring-boot-util/src/main/resources/META-INF/services/org.apache.shardingsphere.driver.spring.boot.datasource.prop.DataSourcePropertiesSetter
@@ -15,6 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.driver.spring.boot.datasource.HikariDataSourcePropertiesSetter
-org.apache.shardingsphere.driver.spring.boot.datasource.CommonDbcp2DataSourcePropertiesSetter
-org.apache.shardingsphere.driver.spring.boot.datasource.TomcatDbcp2DataSourcePropertiesSetter
+org.apache.shardingsphere.driver.spring.boot.datasource.prop.impl.HikariDataSourcePropertiesSetter
+org.apache.shardingsphere.driver.spring.boot.datasource.prop.impl.CommonDbcp2DataSourcePropertiesSetter
+org.apache.shardingsphere.driver.spring.boot.datasource.prop.impl.TomcatDbcp2DataSourcePropertiesSetter
